// src/app/modules/transportation/models/vehicle-pollution-data-view.ts
import { AttachmentViewModel } from 'app/modules/academics/student-admissions/model/attachment-view-model';

/**
 * Transport
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



/**
 * VehicleDataView
 */
export interface VehiclePollutionDataView { 
    /**
     * Gets Or Sets
     */
    vehicleId?: string;
    /**
     * Gets Or Sets
     */
    pollutionCheckDate?: any;
    /**
     * Gets Or Sets
     */
    pollutionCheckExpiryDate?: any;
    /**
     * gets or sets Attachment
     */
    attachment?: AttachmentViewModel;
}